+++ src/msg.rs
@@ -1,0 +1,243 @@
+use cosmwasm_std::Binary;
+use cw721::Expiration;
+use cw721_base::{
+    msg::ExecuteMsg as CW721ExecuteMsg, MintMsg as CW721MintMsg, QueryMsg as CW721QueryMsg,
+};
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug, Default)]
+pub struct Trait {
+    pub display_type: Option<String>,
+    pub trait_type: String,
+    pub value: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug, Default)]
+pub struct Metadata {
+    pub image: Option<String>,
+    pub image_data: Option<String>,
+    pub email: Option<String>,
+    pub external_url: Option<String>,
+    pub public_name: Option<String>,
+    pub public_bio: Option<String>,
+    pub twitter_id: Option<String>,
+    pub discord_id: Option<String>,
+    pub telegram_id: Option<String>,
+    pub keybase_id: Option<String>,
+    pub validator_operator_address: Option<String>,
+}
+
+pub type Extension = Metadata;
+
+pub type MintMsg = CW721MintMsg<Extension>;
+
+// Extended CW721 ExecuteMsg, added the ability to update, burn, and finalize nft
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
+#[serde(rename_all = "snake_case")]
+pub enum ExecuteMsg {
+    /// If the sender is
+    UpdatePreferredAlias {
+        token_id: String,
+    },
+    // Destroys the NFT permanently.
+    Burn {
+        token_id: String,
+    },
+
+    // Set minter (admin)
+    SetMinter {
+        minter: String,
+    },
+
+    /// Mint a new NFT, can only be called by the contract minter
+    Mint(MintMsg),
+
+    // Standard CW721 ExecuteMsg
+    /// Transfer is a base message to move a token to another account without triggering actions
+    TransferNft {
+        recipient: String,
+        token_id: String,
+    },
+    /// Send is a base message to transfer a token to a contract and trigger an action
+    /// on the receiving contract.
+    SendNft {
+        contract: String,
+        token_id: String,
+        msg: Binary,
+    },
+    /// Allows operator to transfer / send the token from the owner's account.
+    /// If expiration is set, then this allowance has a time/height limit
+    Approve {
+        spender: String,
+        token_id: String,
+        expires: Option<Expiration>,
+    },
+    /// Remove previously granted Approval
+    Revoke {
+        spender: String,
+        token_id: String,
+    },
+    /// Allows operator to transfer / send any token from the owner's account.
+    /// If expiration is set, then this allowance has a time/height limit
+    ApproveAll {
+        operator: String,
+        expires: Option<Expiration>,
+    },
+    /// Remove previously granted ApproveAll permission
+    RevokeAll {
+        operator: String,
+    },
+}
+
+impl From<ExecuteMsg> for CW721ExecuteMsg<Extension> {
+    fn from(msg: ExecuteMsg) -> CW721ExecuteMsg<Extension> {
+        match msg {
+            ExecuteMsg::TransferNft {
+                recipient,
+                token_id,
+            } => CW721ExecuteMsg::TransferNft {
+                recipient,
+                token_id,
+            },
+            ExecuteMsg::SendNft {
+                contract,
+                token_id,
+                msg,
+            } => CW721ExecuteMsg::SendNft {
+                contract,
+                token_id,
+                msg,
+            },
+            ExecuteMsg::Approve {
+                spender,
+                token_id,
+                expires,
+            } => CW721ExecuteMsg::Approve {
+                spender,
+                token_id,
+                expires,
+            },
+            ExecuteMsg::Revoke { spender, token_id } => {
+                CW721ExecuteMsg::Revoke { spender, token_id }
+            }
+            ExecuteMsg::ApproveAll { operator, expires } => {
+                CW721ExecuteMsg::ApproveAll { operator, expires }
+            }
+            ExecuteMsg::RevokeAll { operator } => CW721ExecuteMsg::RevokeAll { operator },
+            _ => panic!("cannot covert {:?} to CW721ExecuteMsg", msg),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
+#[serde(rename_all = "snake_case")]
+pub enum QueryMsg {
+    /// Query preferred alias. Takes an address and returns the token id, if set.
+    /// default behaviour is to return the first result, if unset.
+    PreferredAlias { address: String },
+    /// Return the owner of the given token, error if token does not exist
+    /// Return type: OwnerOfResponse
+    OwnerOf {
+        token_id: String,
+        /// unset or false will filter out expired approvals, you must set to true to see them
+        include_expired: Option<bool>,
+    },
+    /// List all operators that can access all of the owner's tokens.
+    /// Return type: `ApprovedForAllResponse`
+    ApprovedForAll {
+        owner: String,
+        /// unset or false will filter out expired approvals, you must set to true to see them
+        include_expired: Option<bool>,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    },
+    /// Total number of tokens issued
+    NumTokens {},
+
+    /// With MetaData Extension.
+    /// Returns top-level metadata about the contract: `ContractInfoResponse`
+    ContractInfo {},
+    /// With MetaData Extension.
+    /// Returns metadata about one particular token, based on *ERC721 Metadata JSON Schema*
+    /// but directly from the contract: `NftInfoResponse`
+    NftInfo { token_id: String },
+    /// With MetaData Extension.
+    /// Returns the result of both `NftInfo` and `OwnerOf` as one query as an optimization
+    /// for clients: `AllNftInfo`
+    AllNftInfo {
+        token_id: String,
+        /// unset or false will filter out expired approvals, you must set to true to see them
+        include_expired: Option<bool>,
+    },
+
+    /// With Enumerable extension.
+    /// Returns all tokens owned by the given address, [] if unset.
+    /// Return type: TokensResponse.
+    Tokens {
+        owner: String,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    },
+    /// With Enumerable extension.
+    /// Requires pagination. Lists all token_ids controlled by the contract.
+    /// Return type: TokensResponse.
+    AllTokens {
+        start_after: Option<String>,
+        limit: Option<u32>,
+    },
+}
+
+impl From<QueryMsg> for CW721QueryMsg {
+    fn from(msg: QueryMsg) -> CW721QueryMsg {
+        match msg {
+            QueryMsg::OwnerOf {
+                token_id,
+                include_expired,
+            } => CW721QueryMsg::OwnerOf {
+                token_id,
+                include_expired,
+            },
+            QueryMsg::ApprovedForAll {
+                owner,
+                include_expired,
+                start_after,
+                limit,
+            } => CW721QueryMsg::ApprovedForAll {
+                owner,
+                include_expired,
+                start_after,
+                limit,
+            },
+            QueryMsg::NumTokens {} => CW721QueryMsg::NumTokens {},
+            QueryMsg::ContractInfo {} => CW721QueryMsg::ContractInfo {},
+            QueryMsg::NftInfo { token_id } => CW721QueryMsg::NftInfo { token_id },
+            QueryMsg::AllNftInfo {
+                token_id,
+                include_expired,
+            } => CW721QueryMsg::AllNftInfo {
+                token_id,
+                include_expired,
+            },
+            QueryMsg::Tokens {
+                owner,
+                start_after,
+                limit,
+            } => CW721QueryMsg::Tokens {
+                owner,
+                start_after,
+                limit,
+            },
+            QueryMsg::AllTokens { start_after, limit } => {
+                CW721QueryMsg::AllTokens { start_after, limit }
+            }
+            _ => panic!("cannot covert {:?} to CW721QueryMsg", msg),
+        }
+    }
+}
+
+// returns a token_id (i.e. a username)
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
+pub struct PreferredAliasResponse {
+    pub username: String,
+}
